requirements.txt
aioredis==1.3.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
cycler==0.10.0
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.7
django-appconf==1.0.4
django-ckeditor==6.0.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-select2==7.4.2
django-storages==1.9.1
entrypoints==0.3
fire==0.1.3
hiredis==1.1.0
httplib2==0.18.1
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server==1.2.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.16.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.3.0
pyrsistent==0.16.0
python-dateutil==2.7.5
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.24.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==16.1.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.1.0
